+++ Model/KlarnaOrder.php
@@ -296,8 +296,15 @@ class KlarnaOrder extends KlarnaOrder_parent
     }
     
     protected function _sendOrderByEmail($oUser = null, $oBasket = null, $oPayment = null) {
-        $oPayment->oxpayments__oxdesc->value = str_replace('Klarna ', '', $oPayment->oxpayments__oxdesc->value);
-        
+
+        if (is_object($oPayment) && in_array($oPayment->oxpayments__oxid->value, KlarnaPayment::getKlarnaPaymentsIds())) {
+            $oPayment->assign(
+                [
+                    'oxdesc' => str_replace('Klarna ', '', $oPayment->getFieldData('oxdesc'))
+                ]
+            );
+        }
+
         return parent::_sendOrderByEmail($oUser, $oBasket, $oPayment);
     }
 }